bpf: Fix ICMP accumulation function to avoid bad checksums.
The icmp_wsum_accumulate() routine uses macros to step over ICMP
payloads in chunks of 128 bytes, but only does this in 256 byte
increments. This means the calculated checksum actually skips
the second half of each increment.

For example, checksum for a 512 byte payload is being calculated,
the function will only include bytes 0-127 and 256-383, with bytes
128-255 and 384-511 skipped. The resulting checksum will be
incorrect.

This has perhaps been overlooked because, in the case of ICMP Error
messages being generated, Cilium only carries the first 64 bytes
(IPv4) or 1232 bytes (IPv16) of the original packet in the error
payload.

Signed-off-by: Alasdair McWilliam <alasdair.mcwilliam@isovalent.com>