ipcache: add null_route flag to IPCache entry structures
Introduce "null_route" flag into IPCache remote_endpoint_info. The intent
is traffic matching an entry with this flag, where the security identity is
classed as WORLD_ID, will be dropped in the datapath.

An example scenario is where Cilium is managing reachability announcements
for an IP Address or Prefix (be that at L2 or L3), such that Cilium will
potentially receive packet carrying unsupported transport protocols. In
some customer scenarios it may be desirable to drop this traffic, rather
than allow it to be punted to the host, because it will likely be forwarded
back to the network.

Using a generic flag with IPCache allows this functionality to be flexible
to multiple other use cases.

Signed-off-by: Alasdair McWilliam <alasdair.mcwilliam@isovalent.com>