pkg/loadbalancer: Sync IPCache entries with Unroutable flag.
This commit introduces logic in the LoadBalancer BPF Reconciler that
upserts and removes entries in the IPCache based on Service resource
definitions using the unsupported-protocol-action annotation.

To be a candidate for programming an Unroutable entry, a Frontend must:

1. be of type ClusterIP, LoadBalancerIP or, ExternalIP.

2. be of an External scope.

3. not be an Unspecified/Zero address, to mitigate risk of
   introducing an IPCache entry that may match inappropriately.

Unlike the wildcard service entry logic, the logic introduced in this
commit is stateless. Frontend-to-IPCacheEntry relationship tracking is
instead provided via the IPCache ResourceID, and is necessary to ensure
we do not remove an IPCache entry that's used by multiple Frontends.
This is called out in commentary in the code itself.

It's also important to highlight that we are programming the IPCache
with metadata originating from source.Generated, which is lower in
priority than other sources such as sources.Kubernetes. This is
important so that a Loadbalancer IPCache entry can be overwritten
by another component. This is also called out in commentary for
clarity.

Signed-off-by: Alasdair McWilliam <alasdair.mcwilliam@isovalent.com>